Study-Plan-React-Backend/requirements.txt
fastapi
uvicorn
langgraph
langchain
langchain_google_genai
python-dotenv
supabase
langchain-core
langchain-community
uvicorn